Lookup Table for Land Cover Change Trajectories

The lookup_trajectory_complete data frame provides a comprehensive lookup table for land cover change trajectories. It maps the combinations of initial and final land cover classes to specific trajectory IDs and defines the corresponding trajectory descriptions.

Usage

lookup_trajectory_complete

Format

A data frame with 64 rows and 7 columns:

ID_1

Numeric identifier for the initial land cover class.

ID_2

Numeric identifier for the final land cover class.

CLASS_1

Description of the initial land cover class.

CLASS_2

Description of the final land cover class.

ID_traj

Unique identifier for each land cover change trajectory, formed by combining ID_1 and ID_2 with an underscore.

trajectory

Detailed description of each land cover change trajectory.

def

Simplified definition or category of each land cover change trajectory.

Source

LUMENSR package

Details

The lookup_trajectory_complete data frame provides a mapping of land cover change trajectories. It considers all possible combinations of initial and final land cover classes and assigns a unique trajectory ID to each combination. The data frame also includes detailed descriptions and simplified definitions for each trajectory.

The land cover classes considered in this lookup table are:

Undisturbed forest

Logged-over forest

Monoculture tree-based plantation

Shrub, grass, and cleared land

Agriculture/annual crop

Mixed tree-based plantation

Others

Settlement and built-up area

The ID_traj column is formed by concatenating the ID_1 and ID_2 values with an underscore, providing a unique identifier for each land cover change trajectory.

The trajectory column provides a detailed description of each land cover change trajectory, indicating the specific transition from the initial land cover class to the final land cover class.

The def column provides a simplified definition or category for each land cover change trajectory, such as "Stable forest," "Reforestation," "Deforestation," or "Other."

Examples

# View the first few rows of the lookup table
head(lookup_trajectory_complete)
#>   ID_1 ID_2                           CLASS_1            CLASS_2 ID_traj
#> 1    1    1                Undisturbed forest Undisturbed forest     1_1
#> 2    2    1                Logged-over forest Undisturbed forest     2_1
#> 3    3    1 Monoculture tree-based plantation Undisturbed forest     3_1
#> 4    4    1    Shrub, grass, and cleared land Undisturbed forest     4_1
#> 5    5    1           Agriculture/annual crop Undisturbed forest     5_1
#> 6    6    1       Mixed tree-based plantation Undisturbed forest     6_1
#>              trajectory           def
#> 1 Stable natural forest Stable forest
#> 2    Recovery to forest Reforestation
#> 3    Recovery to forest Reforestation
#> 4    Recovery to forest Reforestation
#> 5    Recovery to forest Reforestation
#> 6    Recovery to forest Reforestation

# Filter the lookup table for a specific initial land cover class
subset(lookup_trajectory_complete, ID_1 == 1)
#>    ID_1 ID_2            CLASS_1                           CLASS_2 ID_traj
#> 1     1    1 Undisturbed forest                Undisturbed forest     1_1
#> 9     1    2 Undisturbed forest                Logged-over forest     1_2
#> 17    1    3 Undisturbed forest Monoculture tree-based plantation     1_3
#> 25    1    4 Undisturbed forest    Shrub, grass, and cleared land     1_4
#> 33    1    5 Undisturbed forest           Agriculture/annual crop     1_5
#> 41    1    6 Undisturbed forest       Mixed tree-based plantation     1_6
#> 49    1    7 Undisturbed forest                            Others     1_7
#> 57    1    8 Undisturbed forest      Settlement and built-up area     1_8
#>                         trajectory                def
#> 1            Stable natural forest      Stable forest
#> 9       Loss to logged-over forest Forest degradation
#> 17                           Other              Other
#> 25 Loss to bare land and abandoned      Deforestation
#> 33                Loss to cropland      Deforestation
#> 41                           Other              Other
#> 49                           Other              Other
#> 57          Loss to infrastructure      Deforestation